Update environment variable names
We have agreed to use more descriptive environment variable names for clarity across the different applications.
All are prefixed with 'DM_'
Next the name of the application they apply to - 'SUPPLIER_FRONTEND' in this case.
Next a description of it's purpose -e.g. 'AUTH_TOKEN'

So in this case we have DM_SUPPLIER_FRONTEND_API_AUTH_TOKEN.