Send password reset emails with encrypted tokens
Basic flow implemented from "Forgot password" through to password reset.
Emails including a URL with encryped token are sent using the Mandrill API.
The token can be decrypted to find the email address, and the age of the
token can be checked (at present they are only valid for 24 hours from
generation).
This does not yet make any calls to the API to validate that the account
exists, or to update anything.