Use WTForms for forgotten-password form
The forgotten password form was not previously doing any validation
on the email address field - it could be empty or not a valid email
address and still would show the flash message "If this email address
exists we'll send a password reset email".

This adds basic checking that there is something entered and that it looks
like an email address.