Use "reset password" everywhere
Before this change we were using different language across the email
boundary. When requesting to have your password reset we were calling it
"reset password" after clicking the link in the email we were calling it
"change password". This change makes the whole reset password process
use consistent language in the templates, URLs and view names.