Validate that clients count is <= 10
WTForms FieldList raises an AssertionError when called with more
than max_entries. So instead of using it we add a custom form
validator to check that there are less than 10 clients and raise
a validation error otherwise.

This shouldn't normally happen, since there are only 10 inputs on
the page and any additional ones will be dropped when the page is
reloaded.