Make login respect the 'next' parameter
Currently users are always taken to the dashboard page when they log in,
even if the 'next' parameter has been set by Flask when the user tried to
directly access a page that requires login.  In that case it would be nice for
the user to be taken directly to the page they were trying to access after
logging in, rather than having to navigate there from the dashboard.

Doing this exposes a possible security hole as users could potentially be
redirected to a malicious website in the 'next' parameter of a link.
(See https://www.owasp.org/index.php/Unvalidated_Redirects_and_Forwards_Cheat_Sheet).
So to avoid this the 'next_url' validation method checks that the 'next' is a page
that belongs to the app.  If it isn't then the user is taken to the dashboard after
logging in, and the invalid 'next' parameter is ignored.