Sync secure cookie config with admin app
Instead of having our secure cookies set to `False` and then
explicitly set as `True` in a Live environment, it's better to assume
secure cookies and explicitly unset them where needed.
This is how it's been done in the admin app, at any rate.
Also added a test to confirm secure cookies are enabled.