Put validation errors in the right order
Previously if there were errors on both fields the messages were in the wrong order (alphabetical ordering made them the opposite way round from on the page).  By reversing the alphabetical ordering we can guarantee that two errors will be displayed in the same order as the fields on the page.