Integrate with data API
This includes creating a custom wtforms boolean field that supports
being not set. This is because the boolean field in the toolkit has a
radio button for true and one for false. However, it can also be true
that neither of them are checked. This may need to move into utils or
the toolkit.

The form is not confirmed and will likely change.