Service updates: Respond to validation errors returned by the API
This prevents updates with validation errors from being saved, and displays validation errors on the form.

In-line messages are no properly mapped to all possible validation errors - it could be easily done for this small set of fields, but it will be better to harmonise with the messaging mechanism used for SSP flow once it exists.

For now the validation message shown is the last in the list of `validations` for the question in SSP content.