Update breadcrumbs, keep hardcoded framework
Pulled out the breadcrumbs logic from the template into the route.
Talked to @robyoung about adding framework names as parameters into
the route and was convinced that our frontend apps don't need anything
more than a hardcoded framework.  At this point, there's only one
framework we'll accept and we know what that is.