Add tests for specific validation errors
On service section update and draft service section update test that the
error codes from the data API get correctly translated into human
readable errors.

This change also removes the check to see if the posted data evaluates
to True. This is done because the data API does the validation and
therefore if there is a problem with the posted data the API should
raise the error. This also removes the fallthrough condition where the
user is redirected to the edit service page with no message in that
case.

_get_formatted_section_data has been refactored to pull out the type
checking because I found the multi-line if conditions hard to parse.

The class level request patching was removed to make validation error
tests easier but also because it wasn't used by any of the tests.