Prevent questions in other sections from being edited
This change filters out any questions that are not in the current
section. @allait noticed that questions that were not part of the
current section could be edited through the supplier frontend. This
allowed users to change fields that they would otherwise not be allowed
to change such as the lot field.