Add supplier declaration questions
This is a quick and dirty piece of work to get something we can show people.

It takes the newly-added declaration questions from the SSP content and re-uses
the tools we have for editing services to show those questions on a page.

It also pins the SSP content temporarily to master. This is so we can quickly
change the content and redploy the app (maybe on Heroku) to see it update.