Move document questions to separate sections
After looking at different ways to do data + files validation
flow it looks like moving each document upload to a separate
page is the most reasonable option right now.

Mixing documents with other fields on a single form behaves
differently from a simple form:
* If the file is uploaded first but sent to the API together
  with the form fields then, errors in other fields fail the
  form and the file URL is not saved
* If the uploaded file URL is saved to form data and sent
  back after the user resubmits the form then the URL needs
  to be validated (which we don't need to do otherwise).
  If the user closes the page without submitting the form the
  file URL is not saved to the API (which is consistent with
  how other forms behave, but might be different from what
  you'd expect from a file upload field)
* If the file is uploaded and the URL is saved to the API before
  other form fields are sent to the API then the form behaves
  differently from other sections (field value is saved even
  though the form submission showed errors)

The main downside to current approach is that some documents are
optional (SFIA rate card), which means that the whole section can
be skipped, but there's no indication of that on the page.