Save info page as the uploaded draft document URL
Uses a buyer frontend page as a base URL for the document uploads
during the service submission. The URL links to an info page that
displays some text about the state of the document and the reason
it can't be displayed before the framework goes live.

Since the documents will have to be moved once the draft gets a
service ID the URLs will change anyway and temporary base URL
doesn't matter.