Add 'Are you sure' step to service deletion
I couldn't get the Jinja template version of the Notification banner "With some plain text and an action" to work - the button can be rendered OK as the action by adding `| safe` to the action, but it isn't possible to dereference the service name to put in the button text.

So I've gone with the HTML version, which I think is fine.