[#100215328] Add selection questions validation
See: https://www.pivotaltracker.com/story/show/100215328

Add validation for the selection questions. Validations are hard coded
in the frontend. If any validations fail the form will not be submitted
to the API. Once the questions have been successfully submitted the
message on the dashboard will change to "You have made the declaration".

Validations that apply:
- All required fields must be filled in (some are optional, some are
  optional under certain conditions).
- Character limits must not be exceeded
- Formatting rules for DUNS and VAT numbers must be followed