Hide unnamed services from the draft services list
The need for non-stored "New service" name for unnamed services
pops up in different contexts (message notifications, button labels,
copied service names etc.), so in order to avoid these edge cases
we've decided to pretend that unnamed services weren't saved in the
first place.

Service name is the section right after service lot, so we've removed
any links to the draft overview page for unnamed services and split
the description into a separate section to make this step easier.

This way if the user decides not to enter a service name the service
is never displayed in the drafts list and there's no way to get to
any other edit page to enter other service data (apart from manually
entering the URL).

Unnamed drafts can be cleaned up when the framework goes live.