Allow users to not answer checkbox questions
For cetain questions users are supposed to tick all the checkboxes that apply.
It may be the case that none of the checkboxes apply, therefore to not
tick them is a valid response (even though it will bar them from getting on
the framework).

We may, at a later stage, change the validation to give an error if the user
_doesn't_ tick _all_ the boxes.