Refactor tests login method
Refactor the login method method that it only mocks the login
data_api_client for the duration of the login so that other tests on the
same views module are not affected. The mocking of get_user method in a
more aggressive way stays the same because it is needed for subsequent
requests. However, it is now cleaned up in the teardown.

This change also updates all the tests to use the same login method.
Some tests had replaced the core one in an attempt to allow testsing of
the things it was mocking. This should no longer be necesarry.