Fix update-user tests to test update-user
They were pointing at create-user and were not giving any data so the
failure was actually coming from the form validation. I couldn't find
anything useful to test in the response to distinguish them so I've
added an assertion on get_user being called which will at least confirm
that the right view is being requested.