Redirect to service name edit page after making a copy
After discussing this a bit, we've decided to change the way "Make
a copy" works. It still creates a new draft as before, but instead
of returning to the services dashboard it will redirect to the
editing of service name section of the newly created draft.

The assumption is that this will encourage suppliers to give
unique and meaningful names to the copied drafts so that we
don't end up with multiple copies of the same service with the
same name.