Fix "Return to summary" link disappearing after validation errors
[#101571000](https://www.pivotaltracker.com/story/show/101571000)

"Return to summary" was hidden for services without a service name
in 09acc4f. The page rendered after validation only uses the update
data, not the full draft service data, so serviceName should be
explicitly added to the context. This was achieved by adding
serviceName to the `posted_data`, but `posted_data` is no longer
passed to the template as it was replaced by `update_data`.

This adds `serviceName` to update_data for both API and document
validation failures.