[#101692502] Change document upload form display
https://www.pivotaltracker.com/story/show/101692502

Change how the previously uploaded file is displayed on the upload form
so that rather than showing a long URL we show a message saying when the
file was uploaded.

This is implemented in the Jinja2 macro to avoid changing the draft data
in place (and thus lying about the value is) and to avoid doing the
formatting in the frontend toolkit (which doesn't have access to the
required template filters).