More required fields
Following discussions with CCS these fields are now required in the declaration: SQ1-1b, SQ1-1d-i, SQ1-1d-ii, SQ1-1e

SQ1-1b wasn't optional before anyway, but the other three were and this makes them required.

Also fixes a small validation bug - previously using `if not ...` returned true for either `None` or `False`, but the required fields should only be added to the errors for the `False` case, and not for the `None` case.  So using `is False` is the right test.