Only send draft update request to API if any fields were changed
There's an expectation that some suppliers will use the section
editing flows to check filled in questions. When every update
sends a request to the API there's a risk that this would lead
to lost writes if other users are editing the same service at
the same time.

To avoid this we check if any of the submitted fields are different
from existing service data and only send the API update request if
fields were changed.

This works for input fields, but not for empty checkboxes, radios and
file uploads. Since they are not part of the posted data when no value
is selected, they don't trigger the "field change" event, which means
that the request is not sent to the API and there's no "answer required"
error shown.

To keep the "answer required" checks performed by the API we look for
the section questions in the service data and send the posted data to
the API if any of the keys are missing, even if there were no changes
made.