Add CSRF error handler with log in redirect
If a session expires and the first user action is to POST a form
CSRF check fails to find the session token and shows an error.

Since CSRF is triggered before the login_required decorator the user
gets a CSRF error, even though the problem was caused by the expired
session.

To fix this we use a custom CSRF error handler that checks if session
was expired and shows a log in page instead of CSRF error response.

Since CSRF middleware stores a new token in session before error handler
is called, we need to check for `user_id` key instead of checking if the
session is empty.