Remove CSRF token time limit
CSRF tokens are stored in the session cookie, so there's little
security benefit from having an additional expiration time on the
token itself: if the session cookie is leaked the attacker can
generate new CSRF tokens anyway.

Instead, we remove the expiration time on the CSRF token and rely on
session expiration. Once user session is renewed after a fresh log in
a new CSRF token will be generated.