Make content match new order
Some of the messages no longer made sense because they refered to parts of
the process, the order of which has changed.

We also removed some messages which weren't immediately relevant to the user’s
next steps. While the information they conveyed was important, it could be
pushed down to the point at which it was most relevant.