Improve and standardise logs
This is a broad commit that does a few things.

Only put hashes of email addresses in logs. This means that we are not
storing personal information in logs but we can search for log messages
by email address if we hash the address locally first.

Extract variables out into extra parameters. The logging in dmutils has
been changed to attempt to format log messages with the log record
(containing all log variables). This means that if a log message is
provided as a template and the variables provided in the extra dict then
the message will be rendered correctly and the variables will be pulled
out in the JSON logs.