Track when users click buttons
It would be very useful to know, for example, the split between how many users
are clicking the 'save and continue' button and how many are clicking the 'save
and return to service overview'.

This commit adds code similar to the link tracking code to fire an event when
any button is clicked.

By recording the text of the button and the URL of the page that the user is on
we can filter the data to work out what's going on.