Make G7 routes support multi-framework
This change makes the routes that were hard coded to G-Cloud 7 support multiple
frameworks. This change does not include moving any of the routes. This
has been left for a later change as it will break far more tests.

This change has been tested with the functional test suite using both
the @functional-test tag and the @ssp tag.

The list of routes is:
- /frameworks/<framework_slug>
- /frameworks/<framework_slug>/services
- /frameworks/<framework_slug>/declaration
- /frameworks/<framework_slug>/<path:filepath>
- /submission/<framework_slug>/create

The change consisted of a few steps:
- Replace `g-cloud-7` with `<framework_slug>` in the route
- Get the framework details from the data API - this 404s if it doesn't
  exist.
- Pass framework slug to `url_for` calls - some of them have just been hard
  coded as `g-cloud-7` for the moment.
- Update tests to mock `get_framework`.