Add application result letter
Add environment variables for agreements and communications buckets.
Add an environment variable for the base assets URL since the assets
domain is not just for documents.

Build a path for an agreements document based on the framework slug and
the supplier ID of the current user. The agreement documents contain the
supplier ID (for easy referencing later) and also live in a directory
named the supplier ID (for easy browsing of the bucket).

Add a function to get a document URL that works in a similar way to the
draft document URL function but only replaces the host and scheme with
the asset URL if the environment variable is set. This allows the URLs
to work in development even if a reverse proxy is not being used.