Add service submission page with a list of framework lots
Splits framework services into two pages: list of lots and individual
lot pages with lot services.

Both pages are only displayed if framework status is either open or
pending. For pending frameworks page is only displayed to suppliers
who have submitted an application. Feature flag check is removed.

Lot pages check the 'one service limit' flag and:

* Display the list of lot services for lots with multiple services.
  This is similar to the previous submission services page, except
  the services are filtered by lot.

* For one service lots: creates the service for given lot and supplier
  if it doesn't exist and redirects to the service page right away.

Note: creating an empty draft for one-service lots is an action taken
as a result of a GET request (which are not covered by the CSRF
protection), but this is the only approach we have given the current
design of the lots page.