Rewrite draft create page to use the first manifest section
Since draft lot is set as part of the URL the first service submission
page can be different for each lot.

The draft creation endpoint will look up the first editable section for
the given lot and use the submitted form data to create the draft service,
similar to how draft updates work at the moment: data is sent to the API
with a list of page questions to validate required answers, any API errors
returned are displayed on the form page, otherwise the draft is created
and user is redirected to the new draft page.

`create_new_draft_service` calls are updated with the breaking changes
in dmutils 12.0.0.