Order live service listings in reverse framework order
Previously these listed in framework_id order, which is fine when only G5 and G6 are live (G6 (id = 1) appears at the top of the list and G5 (id = 3) below.

But when G-Cloud 7 (id = 4) goes live we want the newest services to be at the top of the list, not at the bottom.

By sorting by reverse framework slug the services will be ordered with the most recent frameworks at the top of the list.