Add edit question section URLs
Specialist questions are edited separately, not as part of the whole
section. In order to support this we use the new method to transform
a multiquestion into a section object. Only questions with `slug` are
supported and the parent section must have `edit_questions` set to
`True` in order to allow individual question edits.