Send virtual page views on declaration completion
Adds a flash message for when the declaration is marked as complete.

Adding method, with tests, to track Virtual Page Views and removing
G-cloud
specific code to make method more generic. This should be moved to the
frontend
toolkit when there is more time.

Add url to flash message to send to Google Analytics for Virtual Page
Views:
```
/suppliers/frameworks/digital-outcomes-and-specialists/declaration_complete
```