Show status of application by lot
Shows the status of a suppliers application to a framework on the page which
lists the lots.

This should work for existing G7 applications, future DOS applications and
any other framework that fits our current model.

The exact messaging depends on:
- whether the lot has a limit of one service
- status of framework
- number of draft and complete services
- status of supplier declaration
- name of the lot

…which makes it quite complicated.

This commit adds a couple of tests to the exiting tests for this endpoint. It
then adds a _lot_ of explicit tests for the functions that generate the
status messages. This is the only way I could think to do this, because the
code itself is not elegant.