Make use of draft service validation errors
After https://github.com/alphagov/digitalmarketplace-api/pull/297
validation errors will come back in API responses when getting a single
draft service. This change makes use of them to decide whether or not to
show the 'mark as complete' button. This is required because not all of
the validation rules can be determined just by whether any required
fields are unfilled.