Jump to summary section after answering a question
It is jarring to be taken back to the top of a long page after answering a
question, and having to scroll all the way down to find where the answer is
played back (and potentially where the next section of questions you want to
answer are).

This commit adds an anchor to jump between these pages more seamlessly.