Add spacing for question heading linebreaks
Added some spacing between linebreaks in our really long question headings.
Only targeting paragraph tags in headings (excluding
the last one), so the rendering for the majority of question
headings doesn't change.

I'm making the change here instead of the frontend toolkit
because the markup in each is different.
The supplier app passes all question headings through a markdown filter
so the `<p>` tags only exist in this app.