Show label for contributor email field
This field was using the `single-question-page`
class. This class currently assumes the page
heading will be a duplicate of the field label so
hides it.

The field was also using a different markup
pattern to other pages with one question (for
example, those in the 'create a new supplier
flow'). The markup for this pattern was missing
some span tags, causing the label to be hidden
normally but shown if validation messages were
present.

This change removes the `single-question-page`
class to ensure the label is shown and swaps the
markup out to use the toolkit textbox pattern
(which has markup that works both for the field
in its base and invalid states).