Change redirect flow for create-users
This commit changes the redirect flow for the /suppliers/create-user
route. Before this change if a logged in user clicked the create user
link for a different email address they would be redirected to the
/suppliers/login route rather than the supplier dashboard. Since changing
the /suppliers/login route to redirect logged in users to the supplier
dashboard this different flow just adds in another redirect step.

With this change we redirect the user to the same place (the supplier
dashboard) but show them an error message telling them what has gone
wrong.

The tests associated with this flow have been updated to not follow
redirects and we instead verify the location header and cookies are set.
The reason for this is that the supplier dashboard has a lot of
dependencies which detract from the meaning of the create user tests.