Pass lot details in to the `edit_submission_section` template
The lot details (name, slug and whether it is limited to a single service) are required in the template so it can properly render breadcrumbs.  Previously the breadcrumbs did not work after a validation error on the page because the lot details weren't included in the service data passed in.