Replace the status-edit radio buttons with a remove button
The new behaviour allows suppliers to make services unavailable
to buy, but not to reactivate them.
As such, the radio buttons that allowed suppliers to change the
status of their service have been replaced (in the case of live
services) with a button to 'remove' them.
Services that have been removed (either by an administrator or
a supplier) just display a message that that have been removed,
with an email address to contact if needed.