Remove status checking from route
When suppliers could change the status of a service back and forth,
additional logic was needed to make sure the submitted status
was legitimate.
Since removing services is now the only option, all of that logic
can be dispensed with and the message we give back to suppliers
can be simplified.

Nor do we need the config that set up the edit status radio buttons.