Use a flash notification for removed services
Previously, removing a service meant passing a lot of
query parameters to the supplier's service listing page.
This wasn't ideal, as hitting 'back' in the browser would
preserve the notification after it was relevant.
Flash messages are how we pass messages everywhere else,
so this is both a better solution and more consistent.