Simplify logic further for viewing/editing a service
Services that aren't published/public can't be modified by
suppliers
and don't need a particularly helpful error message.
This means all of the logic contained at the bottom in a
separate method can be used directly in the `.edit_service`
GET route, as none if it is needed by the `.remove_service`
POST route.