Return "account already exists" message if the user already exists
Removes account type change from buyer to supplier.

Removes the checks for logged in users: if the account already
exists there's nothing we can change about it, so it doesn't matter
if the user is already logged in or not. `/create-user` page won't
log you in after the initial registration, even if you're using the
same token you've used to register, since it's not a general "log in"
route - otherwise it's possible to use it to log in without knowing
the account password.

Changes messages for existing buyer and admin accounts to make
it possible to reuse error the template for buyer registrations.