Ask supplier to confirm removal of service
With other destructive actions we allow (deleting drafts, for
instance), we ask the user to confirm their intent before
doing the thing.
Ported the methodology used for deleting a draft service to
ask the user to confirm removing a draft service.
Ditched the `_remove-service.html` partial template because
it really wasn't giving any useful abstraction.