Change uploaded document names and headers
Because including the supplier names in the filenames for
uploaded documents is quite finicky, we've opted to keep only the
`supplier_id` in the filename proper but suggest a different filename
(which includes the supplier_name) when the file is downloaded.
By setting a `Content-Disposition` header on the uploaded file, we:

- force the browser to download it
- suggest a different filename in the 'Save As' dialog

Note that since we use the same dmutils function for both viewing
and downloading files, and since all existing files are tied to the
old naming convention, we have to update the supplier uploading method
separately from the supplier downloading method.
Once existing files have been renamed, we can port the change over to
dmutils.