Fix if condition to display links properly
After judiciously refactoring the /frameworks/dashboard page, it
turned out that there was a stray if condition checking for a
template variable that was no longer being passed in.

This meant that, after uploading a signed framework agreement and
receiving a countersigned one back, that it would still be possible
to upload a *new* signed framework agreement -- anarchy, basically.

This bugfix removes that possibility.