Load g-cloud-7 service editing manifest
Manifests are loaded initially and then retrieved when they're needed.
This means they'll fail if one tries to retrieve a manifest that wasn't
loaded.
In this case, the `edit_service` manifest for G-Cloud 7 wasn't loaded,
so if any supplier tried to view/remove one of their G7 services, they
would get a 500-level (ContentNotFound) error.
Adding the manifest solves the problem.