Reintroduce removed CSS to framework agreement page
[In a past commit](https://github.com/alphagov/digitalmarketplace-supplier-frontend/commit/dc3a5a24befd38c1bd9738f862c8f60f0f8e71a9),
some CSS was removed that had been ported over to the toolkit.
However, some additional CSS that hadn't made it into the toolkit was
deleted as well, resulting in an unstyled list of instructions on the
framework agreement page.
This commit reintroduces the original CSS for styling the
framework agreement instructions.