Handle case on supplier dashboard if no live date
Frameworks in standstill have a "Live from {date}" subheading on
the supplier dashboard for suppliers who are going to make it onto
the framework.
If there was no date specified, a string containing just "Live from"
would still have been shown.
This commit will hide this subheading if no date is available.