Add route for responding to opportunities
Added a new views file for displaying the brief-response
submission form.  Looks for a live brief from the API,
loads brief-response questions from the content repostory, and
then adds Essential Requirement questions and Nice-to-have
Requirement questions (written by the buyer) to the appropriate
`ContentQuestion`.
Reuses the same template we use to edit services.