Add `convert_to_boolean` filter to Jinja2
Our yes/no questions are returned from the form as strings
and but need to be set to booleans in order for their values
to be reinserted into the form after a validation error.

Easy fix was to reuse the `convert_to_boolean` function that
lives in dmutils.  Registers it as an identically-names Jinja
template.