Add new `boolean_list` macro
Our new `boolean_list` macro differs from all of the others
in that it relies both on static content from the content
repository as well as buyer-submitted string fields which are
unique to each brief.

At some point both sets of data need to be mashed together, but
that could happen in the view, the template, the macro or the
frontend toolkit.

@kevkeenoy and I agreed that, since we're displaying all of the
questions on one page, that putting the supplier-submitted question
data into the `ContentSection` object.  It's not ideal, as there's
a danger of potentially overwriting other keys unintentionally, as
well as the fact that we don't really need the `ContentSection`
object to be any more complicated, but doing it this way allows us to
keep our HTML template totally generic and avoid adding anything
additional to the toolkit.

`boolean_list` has more logic/markup contained within it than our other
macros, but since mixing the user-submitted data with static content is
only necessary at the presentation layer, the macro seemed like a reasonable
choice for storing the complexity.