Implement per-question validation
@tombye rightly pointed out that validating a `boolean_list`
question as a block wouldn't be very useful for someone who'd
missed 1 or 2 essential questions in a long list.  It might
make it hard to see where the actual problem was, and clicking
on the link to the question at the top would jump to the top
of the `boolean_list` and put the focus on the first element.

The solution we went with was to intercept all returned errors
and, for `boolean_list` ones, look for `True`/`False` values
in the response data.  If not found, create a new error in the
format of the existing errors, using specific anchor tags and
questions.

There's an argument that this should go into the content loader,
but as of now, it's only needed in one view in one app.