Move complicated boolean_list logic
Put all the weird stuff we have to do for `boolean_lists`
into the content loader.

Two things, fundamentally:
- Merging the brief-specific questions with the static content
- Getting the correct error messages to return

In order to return the correct error messages, we needed to know
what the questions asked were and what the supplier actually answered.
Passing response data into the `ContentSummary` sets values for
boolean lists if any exist, and then, once the validation
messages are returned from the API, we can see both (a) what
the questions were, and (b) whether they were correctly responded to.

Unfortunately, this means we have to treat boolean_list questions
differently then we treat other ones, but the good news is that
none of the specific changes put in for boolean_lists (I don't think)
have implications for other questions.