Fix supplier view of submissions
Before this change the supplier view of their submissions (the lot
selection page) does not make a lot of sense after the framework is no
longer open. In the case of Digital Outcomes and Specialists the
lot titles were things like 'Apply for Digital Specialists' and link
through to add a service of that type. This doesn't make sense after the
framework is no longer open for submissions.

This is a bit of a hack but it solving it more correctly does not seem
like a priority right now.