Do not send email from unverified domains
Mandrill is going to start preventing emails from unverified domains.
This means that we will no longer be able to send emails with the From
field set to supplier email addresses. This change sets the From field
to a no-reply address on our domain and uses the Reply-To field for the
supplier email address.