Don't show link for DOS services on the supplier dashboard
Since DOS services don't have an edit service manifest or a public
service page on the buyer frontend to link to we don't make the
service name a link for now.

There's no check in the view, so the page would 500 if visited, but
at the same time there's no obvious way to find the DOS service ID
required to manually construct a link.