Add question and answer session details page
Adds a page with brief "webinar" details only visible to eligible
suppliers. The page displays Q and A session details from the brief.

If the details weren't set the link from the brief page should be
hidden, and this view URL will show an empty page with a "Return to
brief" link.